Change raw capacity type to Long
We might have an unsigned 32-bit capacity value soon, and since the JVM
doesn't natively support unsigned numbers, we need to use a Long to
store that.

+++ app/src/main/java/dev/kdrag0n/batterymonitor/data/PsyProperty.kt
@@ -10,5 +10,5 @@ data class PsyProperty(val supply: String, val property: String) {
     }
 
     fun exists() = file.exists()
-    fun read() = file.readText().toInt()
+    fun read() = file.readText().toLong()
 }
